[errormap] Move error maps into bytecode metadata. (#3220)
This moves the error explanation map (that info which allows to display reasons like `ESOME_REASON` and the related comment on Move Abort) into the metadata section of the Move bytecode and let it be interpreted by the Aptos VM. This is necessary as when we upgrade the framework, the error map  compiled into the binary becomes outdated. Also, this enables the same mechanism for user code.

There is now an optional `AbortInfo` embedded into an `ExecutionStatus` which the VM populates from the metadata. The overhead this creates should be marginal, and the deserialized metadata is cached on VM level.

A test in `e2e-move-tests/tests/error_map.rs` shows that it works.

As a consequence, explicit representation of errors maps in package metadata or elsewhere is not longer needed and has been removed.

Also the existing `error_explain` is not longer needed. This made the `sdk/aptos_transaction_builder` crate empty, and so I removed it. The crate served to re-export `cached-packages`, which let to some cascading changes. In general we should avoiding re-exports in our internal code base because it makes refactoring and reasoning about code harder. (It's ok to re-export for external APIs. If we need `aptos-transaction-builder` for this purpose we can re-introduce it, but it should not be used by our internal code.)